Simplify ticker service
