+++ app/src/main/java/com/alphawallet/app/entity/Ticker.java
@@ -1,5 +1,6 @@
 package com.alphawallet.app.entity;
 
+import com.alphawallet.app.entity.tokens.TokenTicker;
 import com.google.gson.annotations.SerializedName;
 
 public class Ticker {
@@ -10,4 +11,19 @@ public class Ticker {
     @SerializedName("percent_change_24h")
     public String percentChange24h;
     public String price_usd;
+
+    public Ticker()
+    {
+
+    }
+
+    public Ticker(TokenTicker tokenTicker, String sym)
+    {
+        id = tokenTicker.id;
+        name = tokenTicker.contract;
+        symbol = sym;
+        price = tokenTicker.price;
+        percentChange24h = tokenTicker.percentChange24h;
+        price_usd = tokenTicker.price;
+    }
 }
